requirements.txt
dbt-bigquery
dlt
dlt[gs]
flatten-json
flatsplode
gcsfs
lancedb
modal
orjson>=3.9.15
polars
pyarrow
pydantic[email]
python-dotenv
pydantic>=2.4.0
reflex
s3fs
uuid7
openai
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
